+++ command/agent/http_test.go
@@ -38,7 +38,7 @@ func makeHTTPServerWithConfig(t *testing.T, cb func(c *Config)) (string, *HTTPSe
 		t.Fatalf("err: %v", err)
 	}
 	conf.UiDir = uiDir
-	servers, err := NewHTTPServers(agent, conf, nil, agent.logOutput)
+	servers, err := NewHTTPServers(agent, conf, agent.logOutput)
 	if err != nil {
 		t.Fatalf("err: %v", err)
 	}
@@ -148,7 +148,7 @@ func TestHTTPServer_UnixSocket_FileExists(t *testing.T) {
 	defer os.RemoveAll(dir)
 
 	// Try to start the server with the same path anyways.
-	if _, err := NewHTTPServers(agent, conf, nil, agent.logOutput); err != nil {
+	if _, err := NewHTTPServers(agent, conf, agent.logOutput); err != nil {
 		t.Fatalf("err: %s", err)
 	}
 
@@ -514,6 +514,39 @@ func TestACLResolution(t *testing.T) {
 			t.Fatalf("bad: %s", token)
 		}
 	})
+}
+
+func TestScadaHTTP(t *testing.T) {
+	// Create the agent
+	dir, agent := makeAgent(t, nextConfig())
+	defer os.RemoveAll(dir)
+	defer agent.Shutdown()
+
+	// Create a generic listener
+	list, err := net.Listen("tcp", ":0")
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+	defer list.Close()
+
+	// Create the SCADA HTTP server
+	scadaHttp := newScadaHttp(agent, list)
+
+	// Returned server uses the listener and scada addr
+	if scadaHttp.listener != list {
+		t.Fatalf("bad listener: %#v", scadaHttp)
+	}
+	if scadaHttp.addr != scadaHTTPAddr {
+		t.Fatalf("expected %v, got: %v", scadaHttp.addr, scadaHTTPAddr)
+	}
+
+	// Check that debug endpoints were not enabled. This will cause
+	// the serve mux to panic if the routes are already handled.
+	mockFn := func(w http.ResponseWriter, r *http.Request) {}
+	scadaHttp.mux.HandleFunc("/debug/pprof/", mockFn)
+	scadaHttp.mux.HandleFunc("/debug/pprof/cmdline", mockFn)
+	scadaHttp.mux.HandleFunc("/debug/pprof/profile", mockFn)
+	scadaHttp.mux.HandleFunc("/debug/pprof/symbol", mockFn)
 }
 
 // assertIndex tests that X-Consul-Index is set and non-zero